게시글 기본 기능 구현 완료

+++ findi_back/articles/views.py
@@ -1,5 +1,5 @@
 from rest_framework.decorators import api_view, permission_classes
-from rest_framework.permissions import IsAuthenticatedOrReadOnly
+from rest_framework.permissions import IsAuthenticatedOrReadOnly, IsAuthenticated
 from rest_framework.response import Response
 from rest_framework import status
 from django.shortcuts import get_object_or_404
@@ -16,15 +16,16 @@ def article_list_create(request):
             articles = Article.objects.filter(title__icontains=search).order_by('-created_at')
         else:
             articles = Article.objects.all().order_by('-created_at')
-        serializer = ArticleSerializer(articles, many=True)
+        serializer = ArticleSerializer(articles, many=True, context={'request': request})
         return Response(serializer.data)
 
     elif request.method == 'POST':
-        serializer = ArticleSerializer(data=request.data)
+        serializer = ArticleSerializer(data=request.data, context={'request': request})
         if serializer.is_valid():
             serializer.save(user=request.user)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
 
 # 게시글 상세 조회, 수정 및 삭제
 @api_view(['GET', 'PUT', 'DELETE'])
@@ -33,14 +34,14 @@ def article_detail(request, pk):
     article = get_object_or_404(Article, pk=pk)
 
     if request.method == 'GET':
-        serializer = ArticleSerializer(article)
+        serializer = ArticleSerializer(article, context={'request': request})
         return Response(serializer.data)
 
     if request.user != article.user:
         return Response({'detail': '권한이 없습니다.'}, status=status.HTTP_403_FORBIDDEN)
 
     if request.method == 'PUT':
-        serializer = ArticleSerializer(article, data=request.data, partial=True)
+        serializer = ArticleSerializer(article, data=request.data, partial=True, context={'request': request})
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
@@ -49,6 +50,7 @@ def article_detail(request, pk):
     elif request.method == 'DELETE':
         article.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
+
 
 # 조회수 증가 로직
 @api_view(['POST'])
@@ -59,6 +61,7 @@ def increment_views(request, pk):
     article.save()
     return Response({'views': article.views})
 
+
 # 댓글 리스트 및 생성
 @api_view(['GET', 'POST'])
 @permission_classes([IsAuthenticatedOrReadOnly])
@@ -67,39 +70,35 @@ def comment_list_create(request, article_id):
 
     if request.method == 'GET':
         comments = article.comments.all().order_by('created_at')
-        serializer = CommentSerializer(comments, many=True)
+        serializer = CommentSerializer(comments, many=True, context={'request': request})
         return Response(serializer.data)
 
     elif request.method == 'POST':
-        serializer = CommentSerializer(data=request.data)
+        serializer = CommentSerializer(data=request.data, context={'request': request})
         if serializer.is_valid():
             serializer.save(user=request.user, article=article)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-# views.py
-@api_view(['PUT'])
-@permission_classes([IsAuthenticatedOrReadOnly])
-def comment_update(request, article_id, comment_id):
+
+
+# 댓글 수정 및 삭제
+@api_view(['PUT', 'DELETE'])
+@permission_classes([IsAuthenticated])
+def comment_update_or_delete(request, article_id, comment_id):
     comment = get_object_or_404(Comment, id=comment_id, article_id=article_id)
 
     if request.user != comment.user:
         return Response({'detail': '권한이 없습니다.'}, status=status.HTTP_403_FORBIDDEN)
 
-    serializer = CommentSerializer(comment, data=request.data, partial=True)
-    if serializer.is_valid():
-        serializer.save()
-        return Response(serializer.data)
-    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    if request.method == 'PUT':
+        serializer = CommentSerializer(comment, data=request.data, partial=True, context={'request': request})
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-# 댓글 삭제
-@api_view(['DELETE'])
-@permission_classes([IsAuthenticatedOrReadOnly])
-def comment_delete(request, article_id, comment_id):
-    comment = get_object_or_404(Comment, id=comment_id, article_id=article_id)
+    elif request.method == 'DELETE':
+        comment.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
 
-    if request.user != comment.user:
-        return Response({'detail': '권한이 없습니다.'}, status=status.HTTP_403_FORBIDDEN)
-
-    comment.delete()
-    return Response(status=status.HTTP_204_NO_CONTENT)
